Vulkan doesn't need to call Win32 SwapBuffers
