+++ windowing/win32.go
@@ -9,14 +9,6 @@ import (
 
 /*
 #include "windowing.h"
-#include <windows.h>
-#include <windowsx.h>
-
-void window_swap_buffers(void* handle) {
-	HWND hwnd = (HWND)handle;
-	HDC hdc = GetDC(hwnd);
-	SwapBuffers(hdc);
-}
 */
 import "C"
 
@@ -59,8 +51,4 @@ func createWindow(windowName string, width, height int, evtSharedMem *evtMem) {
 	windowTitle := utf16.Encode([]rune(windowName))
 	title := (*C.wchar_t)(unsafe.Pointer(&windowTitle[0]))
 	go C.window_main(title, C.int(width), C.int(height), evtSharedMem.AsPointer(), evtSharedMemSize)
-}
-
-func swapBuffers(handle unsafe.Pointer) {
-	C.window_swap_buffers(handle)
 }
