+++ windowing/window.vk.go
@@ -2,7 +2,10 @@
 
 package windowing
 
-import "kaiju/rendering"
+import (
+	"kaiju/rendering"
+	"unsafe"
+)
 
 func selectRenderer(w *Window, name string) (rendering.Renderer, error) {
 	return rendering.NewVKRenderer(w, name)
@@ -15,4 +18,7 @@ func (w *Window) GetDrawableSize() (int32, int32) {
 func (w *Window) GetInstanceExtensions() []string {
 	// TODO:  VK_KHR_win32_surface is windows specific
 	return []string{"VK_KHR_surface\x00", "VK_KHR_win32_surface\x00"}
+}
+
+func swapBuffers(handle unsafe.Pointer) {
 }
